src/example/requirements.txt
configobj==5.0.6
grpcio==1.26.0
grpcio-reflection==1.22.0
grpcio-tools==1.26.0
protobuf==3.11.2
six==1.14.0